# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang engineer test assessment - the Spy Cat Agency",
        "title": "Golang engineer test assessment - the Spy Cat Agency",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/v1",
    "paths": {
...
        "/mission": {
            "post": {
                "description": "Create new Mission",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mission"
                ],
                "summary": "Create Mission",
                "parameters": [
                    {
                        "description": "Mission payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.MissionPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.MissionWithTargets"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/mission/mission_list": {
            "get": {
                "description": "List of missions",
                "tags": [
                    "mission"
                ],
                "summary": "List of missions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.MissionWithMetadata"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/mission/{id}": {
            "get": {
                "description": "Get one of mission by ID",
                "tags": [
                    "mission"
                ],
                "summary": "Get one of mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Delete mission by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mission"
                ],
                "summary": "Delete mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Update mission by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mission"
                ],
                "summary": "Update mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.UpdatedMission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/mission/{id}/cat/{cat_id}": {
            "patch": {
                "description": "Add Spy Cat to Mission",
                "tags": [
                    "mission"
                ],
                "summary": "Add Spy Cat to Mission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Mission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cat ID",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.MissionPayload": {
            "type": "object",
            "required": [
                "complete",
                "targets"
            ],
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "targets": {
                    "type": "array",
                    "maxItems": 3,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/main.Target"
                    }
                }
            }
        },
        "main.Target": {
            "type": "object",
            "required": [
                "complete",
                "country",
                "name",
                "notes"
            ],
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "notes": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "store.Cat": {
            "type": "object",
            "properties": {
                "breed": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "year_of_experience": {
                    "type": "integer"
                }
            }
        },
        "store.Mission": {
            "type": "object",
            "properties": {
                "cat_id": {
                    "type": "integer"
                },
                "completed": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "store.MissionWithMetadata": {
            "type": "object",
            "properties": {
                "cat": {
                    "$ref": "#/definitions/store.Cat"
                },
                "mission": {
                    "$ref": "#/definitions/store.Mission"
                }
            }
        },
        "store.MissionWithTargets": {
            "type": "object",
            "properties": {
                "mission": {
                    "$ref": "#/definitions/store.Mission"
                },
                "targets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.Target"
                    }
                }
            }
        },
        "store.Target": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mission_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                }
            }
        },
...
        "store.UpdatedMission": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}